requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
motor==3.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
pymongo==4.6.1
qrcode==7.4.2
pillow==10.3.0
openpyxl==3.1.2
jinja2==3.1.3
python-multipart==0.0.18
email-validator==2.2.0
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
